backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
aiomysql
pydantic-settings
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
python-multipart
slowapi
pytest
httpx
